pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["Bulat <bulat.kajrov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
celery = {extras = ["redis"], version = "^5.3.1"}
django-allauth = "^0.54.0"
django-debug-toolbar = "^4.2.0"
django-redis = "^5.3.0"
flake8 = "^6.1.0"
gunicorn = "^21.2.0"
isort = "^5.12.0"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.2"
djoser = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
